# dist/api-docs.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "eCommerce API",
      "description": "Documentation for Ecommerce API."
    },
    "schemes": [
      "http" 
    ],
    "host": "",
    "consumes": [
      "application/json",
      "application/x-www-form-urlencoded",
      "multipart/form-data"
    ],
...
    "definitions": {
      "Department": {
        "properties": {
          "department_id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Regional"
          },
          "description": {
            "type": "string",
            "example": "Proud of your country? Wear a T-shirt with a national symbol stamp!"
          }
        }
      },
      "Category": {
        "properties": {
          "category_id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "French"
          },
          "description": {
            "type": "string",
            "example": "The French have always had an eye for beauty. One look at the T-shirts below and you'll see that same appreciation has been applied abundantly to their postage stamps. Below are some of our most beautiful and colorful T-shirts, so browse away! And don't forget to go all the way to the bottom - you don't want to miss any of them!"
          },
          "department_id": {
            "type": "integer",
            "example": 1
          }
        }
      },
...
      "Error": {
        "properties": {
          "code": {
            "type": "string",
            "example": "USR_02"
          },
          "message": {
            "type": "string",
            "example": "The field example is empty."
          },
          "field": {
            "type": "string",
            "example": "example"
          },
          "status": {
            "type": "string",
            "example": "500"
          }
        }
      },
...
    },
    "paths": {
      "/departments": {
        "get": {
          "tags": [
            "departments"
          ],
          "summary": "Get Departments",
          "description": "Return a list of department.",
          "responses": {
            "200": {
              "description": "A Array of Object Department",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Department"
                }
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "security": []
        }
      },
      "/departments/{department_id}": {
        "get": {
          "tags": [
            "departments"
          ],
          "summary": "Get Department by ID",
          "description": "Return a department by ID.",
          "parameters": [
            {
              "name": "department_id",
              "in": "path",
              "required": true,
              "description": "ID of Department",
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "A object of Department",
              "schema": {
                "$ref": "#/definitions/Department"
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "security": []
        }
      },
      "/tax": {
        "get": {
          "tags": [
            "tax"
          ],
          "summary": "Get All Taxes",
          "description": "Return a list of tax.",
          "responses": {
            "200": {
              "description": "A Array of Object Tax",
              "schema": {
                "type": "array",
                "items": {
                  "properties": {
                    "tax_id": {
                      "type": "integer",
                      "example": 1
                    },
                    "tax_type": {
                      "type": "string",
                      "example": "Sales Tax at 8.5%"
                    },
                    "tax_percentage": {
                      "type": "string",
                      "example": "8.50"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "security": []
        }
      },
      "/tax/{tax_id}": {
        "get": {
          "tags": [
            "tax"
          ],
          "summary": "Get Tax by ID",
          "description": "Return a tax by ID.",
          "parameters": [
            {
              "name": "tax_id",
              "in": "path",
              "required": true,
              "description": "ID of Tax",
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "A object of Tax",
              "schema": {
                "properties": {
                  "tax_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "tax_type": {
                    "type": "string",
                    "example": "Sales Tax at 8.5%"
                  },
                  "tax_percentage": {
                    "type": "string",
                    "example": "8.50"
                  }
                }
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "security": []
        }
      },
      "/categories": {
        "get": {
          "tags": [
            "categories"
          ],
          "summary": "Get Categories",
          "description": "Return a list of categories.",
          "parameters": [
            {
              "name": "order",
              "in": "query",
              "description": "Sorting a field. Allowed fields: 'category_id', 'name'.",
              "required": false,
              "type": "string"
            },
            {
              "name": "page",
              "in": "query",
              "description": "Inform the page. Starting with 1. Default: 1",
              "required": false,
              "type": "integer"
            },
            {
              "name": "limit",
              "in": "query",
              "description": "Limit per page, Default: 20.",
              "required": false,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Return a list with count (total categories) and the rows of Categories",
              "schema": {
                "properties": {
                  "count": {
                    "type": "integer",
                    "example": 40
                  },
                  "rows": {
                    "type": "array",
                    "items": {
                      "$ref": "#/definitions/Category"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "security": []
        }
      },
      "/categories/{category_id}": {
        "get": {
          "tags": [
            "categories"
          ],
          "summary": "Get Category by ID",
          "description": "Return a category by ID.",
          "parameters": [
            {
              "name": "category_id",
              "in": "path",
              "required": true,
              "description": "ID of Category",
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Return a object of Category",
              "schema": {
                "$ref": "#/definitions/Category"
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "security": []
        }
      },
      "/categories/inProduct/{product_id}": {
        "get": {
          "tags": [
            "categories"
          ],
          "summary": "Get Categories of a Product",
          "description": "Return a list of categories from a Product ID",
          "responses": {
            "200": {
              "description": "Return a array of Category Objects",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/CategoryBasic"
                }
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "parameters": [
            {
              "name": "product_id",
              "in": "path",
              "required": true,
              "description": "ID of Product",
              "type": "integer"
            }
          ],
          "security": []
        }
      },
      "/categories/inDepartment/{department_id}": {
        "get": {
          "tags": [
            "categories"
          ],
          "summary": "Get Categories of a Department",
          "description": "Return a list of categories from a Department ID",
          "responses": {
            "200": {
              "description": "Return a array of Object Category",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Category"
                }
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "parameters": [
            {
              "name": "department_id",
              "in": "path",
              "required": true,
              "description": "ID of Department",
              "type": "integer"
            }
          ],
          "security": []
        }
      },
      "/products": {
        "get": {
          "tags": [
            "products"
          ],
          "summary": "Get All Products",
          "description": "Return a list of products.",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "description": "Inform the page. Starting with 1. Default: 1",
              "required": false,
              "type": "integer"
            },
            {
              "name": "limit",
              "in": "query",
              "description": "Limit per page, Default: 20.",
              "required": false,
              "type": "integer"
            },
            {
              "name": "description_length",
              "in": "query",
              "description": "Limit of the description, Default: 200.",
              "required": false,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Return the total of products and a list of Products in row.",
              "schema": {
                "properties": {
                  "count": {
                    "type": "integer",
                    "example": 40
                  },
                  "rows": {
                    "type": "array",
                    "items": {
                      "$ref": "#/definitions/Product"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "security": []
        }
      },
      "/products/search": {
        "get": {
          "tags": [
            "products"
          ],
          "summary": "Search products",
          "responses": {
            "200": {
              "description": "Return the total of products and a list of products.",
              "schema": {
                "properties": {
                  "count": {
                    "type": "integer",
                    "example": 40
                  },
                  "rows": {
                    "type": "array",
                    "items": {
                      "$ref": "#/definitions/Product"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "parameters": [
            {
              "name": "query_string",
              "in": "query",
              "description": "Query to search.",
              "required": true,
              "type": "string"
            },
            {
              "name": "all_words",
              "in": "query",
              "description": "All words or no. Default: on",
              "required": false,
              "type": "string",
              "enum": [
                "on",
                "off"
              ]
            },
            {
              "name": "page",
              "in": "query",
              "description": "Inform the page. Starting with 1. Default: 1",
              "required": false,
              "type": "integer"
            },
            {
              "name": "limit",
              "in": "query",
              "description": "Limit per page, Default: 20.",
              "required": false,
              "type": "integer"
            },
            {
              "name": "description_length",
              "in": "query",
              "description": "Limit of the description, Default: 200.",
              "required": false,
              "type": "integer"
            }
          ],
          "security": []
        }
      },
      "/products/{product_id}": {
        "get": {
          "tags": [
            "products"
          ],
          "summary": "Product by ID",
          "responses": {
            "200": {
              "description": "Return a Product Object",
              "schema": {
                "$ref": "#/definitions/ProductComplete"
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "parameters": [
            {
              "name": "product_id",
              "in": "path",
              "description": "Product ID",
              "required": true,
              "type": "integer"
            }
          ],
          "security": []
        }
      },
      "/products/inCategory/{category_id}": {
        "get": {
          "tags": [
            "products"
          ],
          "summary": "Get a lit of Products of Categories",
          "responses": {
            "200": {
              "description": "Return a list of Product Objects",
              "schema": {
                "properties": {
                  "count": {
                    "type": "integer",
                    "example": 40
                  },
                  "rows": {
                    "type": "array",
                    "items": {
                      "$ref": "#/definitions/Product"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "parameters": [
            {
              "name": "category_id",
              "in": "path",
              "description": "Category ID",
              "required": true,
              "type": "integer"
            },
            {
              "name": "page",
              "in": "query",
              "description": "Inform the page. Starting with 1.",
              "required": false,
              "type": "integer"
            },
            {
              "name": "limit",
              "in": "query",
              "description": "Limit per page.",
              "required": false,
              "type": "integer"
            },
            {
              "name": "description_length",
              "in": "query",
              "description": "Limit to description.",
              "required": false,
              "type": "integer"
            }
          ],
          "security": []
        }
      },
      "/products/inDepartment/{department_id}": {
        "get": {
          "tags": [
            "products"
          ],
          "summary": "Get a list of Products on Department",
          "responses": {
            "200": {
              "description": "Return the total and a list of products",
              "schema": {
                "properties": {
                  "count": {
                    "type": "integer",
                    "example": 40
                  },
                  "rows": {
                    "type": "array",
                    "items": {
                      "$ref": "#/definitions/ProductInDepartment"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "parameters": [
            {
              "name": "department_id",
              "in": "path",
              "description": "Department ID",
              "required": true,
              "type": "integer"
            },
            {
              "name": "page",
              "in": "query",
              "description": "Inform the page. Starting with 1.",
              "required": false,
              "type": "integer"
            },
            {
              "name": "limit",
              "in": "query",
              "description": "Limit per page.",
              "required": false,
              "type": "integer"
            },
            {
              "name": "description_length",
              "in": "query",
              "description": "Limit to description.",
              "required": false,
              "type": "integer"
            }
          ],
          "security": []
        }
      },
      "/products/{product_id}/details": {
        "get": {
          "tags": [
            "products"
          ],
          "summary": "Get details of a Product",
          "responses": {
            "200": {
              "description": "Return a Object of Products",
              "schema": {
                "$ref": "#/definitions/ProductDetail"
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "parameters": [
            {
              "name": "product_id",
              "in": "path",
              "description": "Product ID",
              "required": true,
              "type": "integer"
            }
          ],
          "security": []
        }
      },
      "/products/{product_id}/locations": {
        "get": {
          "tags": [
            "products"
          ],
          "summary": "Get locations of a Product",
          "responses": {
            "200": {
              "description": "Return locations of products.",
              "schema": {
                "$ref": "#/definitions/ProductLocations"
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "parameters": [
            {
              "name": "product_id",
              "in": "path",
              "description": "Product ID",
              "required": true,
              "type": "integer"
            }
          ],
          "security": []
        }
      },
      "/products/{product_id}/reviews": {
        "get": {
          "tags": [
            "products"
          ],
          "summary": "Get reviews of a Product",
          "responses": {
            "200": {
              "description": "Return a list of reviews",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Review"
                }
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "parameters": [
            {
              "name": "product_id",
              "in": "path",
              "description": "Product ID",
              "required": true,
              "type": "integer"
            }
          ],
          "security": []
        },
        "post": {
          "tags": [
            "products"
          ],
          "responses": {
            "200": {
              "description": "No data."
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "401": {
              "description": "Unauthorized",
              "schema": {
                "$ref": "#/definitions/Unauthorized"
              }
            }
          },
          "parameters": [
            {
              "name": "product_id",
              "in": "path",
              "description": "Product ID",
              "required": true,
              "type": "integer"
            },
            {
              "name": "review",
              "in": "formData",
              "description": "Review Text of Product",
              "required": true,
              "type": "string"
            },
            {
              "name": "rating",
              "in": "formData",
              "description": "Rating of Product",
              "required": true,
              "type": "integer"
            }
          ],
          "security": [
            {
              "UserSecurity": []
            }
          ]
        }
      },
      "/customer": {
        "put": {
          "tags": [
            "customers"
          ],
          "summary": "Update a customer",
          "responses": {
            "200": {
              "description": "A Customer Object",
              "schema": {
                "$ref": "#/definitions/Customer"
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "401": {
              "description": "Unauthorized",
              "schema": {
                "$ref": "#/definitions/Unauthorized"
              }
            }
          },
          "parameters": [
            {
              "name": "name",
              "in": "formData",
              "description": "Customer name.",
              "required": true,
              "type": "string"
            },
            {
              "name": "email",
              "in": "formData",
              "description": "Customer email.",
              "required": true,
              "type": "string"
            },
            {
              "name": "password",
              "in": "formData",
              "description": "Customer password.",
              "required": false,
              "type": "string"
            },
            {
              "name": "day_phone",
              "in": "formData",
              "description": "Customer day phone.",
              "required": false,
              "type": "string"
            },
            {
              "name": "eve_phone",
              "in": "formData",
              "description": "Customer eve phone.",
              "required": false,
              "type": "string"
            },
            {
              "name": "mob_phone",
              "in": "formData",
              "description": "Customer mob phone.",
              "required": false,
              "type": "string"
            }
          ],
          "security": [
            {
              "UserSecurity": []
            }
          ]
        },
        "get": {
          "tags": [
            "customers"
          ],
          "summary": "Get a customer by ID. The customer is getting by Token.",
          "responses": {
            "200": {
              "description": "Return a Object of Customer with auth credencials",
              "schema": {
                "$ref": "#/definitions/Customer"
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            },
            "401": {
              "description": "Unauthorized",
              "schema": {
                "$ref": "#/definitions/Unauthorized"
              }
            }
          },
          "security": [
            {
              "UserSecurity": []
            }
          ]
        }
      },
      "/customers": {
        "post": {
          "tags": [
            "customers"
          ],
          "summary": "Register a Customer",
          "description": "Record a customer.",
          "responses": {
            "200": {
              "description": "Return a Object of Customer with auth credencials",
              "schema": {
                "$ref": "#/definitions/CustomerRegister"
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "parameters": [
            {
              "name": "name",
              "in": "formData",
              "description": "Name of User.",
              "required": true,
              "type": "string"
            },
            {
              "name": "email",
              "in": "formData",
              "description": "Email of User.",
              "required": true,
              "type": "string"
            },
            {
              "name": "password",
              "in": "formData",
              "description": "Password of User.",
              "required": true,
              "type": "string"
            }
          ],
          "security": []
        }
      },
      "/customers/login": {
        "post": {
          "tags": [
            "customers"
          ],
          "summary": "Sign in in the Shopping.",
          "description": "Customer Login.",
          "responses": {
            "200": {
              "description": "Return a Object of Customer with auth credencials",
              "schema": {
                "$ref": "#/definitions/CustomerRegister"
              }
            },
            "400": {
              "description": "Return a error object",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          },
          "parameters": [
            {
              "name": "email",
              "in": "formData",
              "description": "Email of User.",
              "required": true,
              "type": "string"
            },
            {
              "name": "password",
              "in": "formData",
              "description": "Password of User.",
              "required": true,
              "type": "string"
            }
          ],
          "security": []
        }
      },
...
    },
    "securityDefinitions": {
      "UserSecurity": {
        "type": "apiKey",
        "in": "header",
        "name": "USER-KEY"
      }
    }
  }